add simple disassemble example
Earlier today I had a file on disk which _should_ have been raw RISCV
machine code, but I wasn't sure. I thought that it would be easy to feed
it into a disassembler to confirm, but, surprisingly, it seems that
there are no RISCV disassemblers that work with raw machine code?
Everything wants to start with elf, but I don't have an elf, I have a
raw pile of bytes! So, surpsing as it is, the fastest way to do what I
wanted was to use this crate to write a short disassembler?

I figured I might as well PR this --- maybe someone else finds it useful!